Moved out image mode tests
